Fix bug that makes plug crash when query fails
PR#10 introduced Query to push out real Error objects to the user in case
the request fails. Dependent on where the request fails, request (here the
module) pushes null for the res object. When creating the error object,
processEntry didn't take care of checking for the res object to exist
before accessing one of its members (in this case statusCode).
The added check will fix the problem.