add optional node-style callbacks to login() and connect()
This should make simple usage quite a lot easier. Previously the only
way to log in looked like:

    plugged.login(myCredentials);
    plugged.on(plugged.LOGIN_SUCCESS, function () {
        // do all my stuff
    });
    plugged.on(plugged.LOGIN_ERROR, function (err) {
        // deal with err
    });

Whereas the usual, and now possible, node way looks more like:

    plugged.login(myCredentials, function (err) {
        if(err) {
             // deal with err
        } else {
             // do all my stuff
        }
    });

As an added bonus, the login() and connect() methods can now be
passed to libraries like Async.js or be automatically promisified in
Bluebird and friends.