Replace iterator factory with class
Replaces the iterator factory createIterator with the class Iterator.
The class approach is more readable as well as getting rid of a silent bug
that caused the literals to access the same currentIndex.

Waterfall has the possibility to define a context in which the queued
functions should run instead of binding them directly.