Remove unnecessary conditional
one check for res was unnecessary since the variable is ensured to exist
at given time, other changes include the change of the default value of
watcherID to -1 as well as setting it as a Timeout instead of an Interval.
The watcher will now stop to run once *_process* indicates that the max
count of parallel connections has been reached.