Add documentation to repo
since the old documentation was fragmented and hard to follow
I decided to put together a new one with a lot more of content
which also gets checked back with code on every change.
The major difference is that this one has a much better structure
to follow as well as being way easier to update.