Remove events JOINED_ROOM and PLUG_ERROR
this was a major flaw in the general design of plugged. guest and connect are really context related and shouldn't be emitting events like a general purpose handler. This really restricted the user in keeping it context based and also blew up the handling code for whether to use a callback or not.