Fix bug which causes an flood api event
There was an edge case in which, when called with a certain timing,
the offset would be reset so that the query would be worked off with zero offset.
Thanks to this fix there's a 400 ms buffer before clearing the offset to make sure newly received requests
are not accidentally processed directly.